requirements.txt
fastapi
ultralytics
opencv-python
numpy
nbformat
pillow
pydantic
python-dotenv
python-multipart
requests
uvicorn
